# `ultralytics 8.3.32` Always transfer `Conv` layer pretrained weights