--- a/ultralytics/nn/tasks.py
+++ b/ultralytics/nn/tasks.py
@@ -281,8 +281,12 @@ class BaseModel(torch.nn.Module):
         """
         model = weights["model"] if isinstance(weights, dict) else weights  # torchvision models are not dicts
         csd = model.float().state_dict()  # checkpoint state_dict as FP32
-        csd = intersect_dicts(csd, self.state_dict())  # intersect
-        self.load_state_dict(csd, strict=False)  # load
+        updated_csd = intersect_dicts(csd, self.state_dict())  # intersect
+        self.load_state_dict(updated_csd, strict=False)  # load
+        first_conv = "model.0.conv.weight"
+        if first_conv not in updated_csd:
+            channel = min(self.state_dict()[first_conv].shape[1], csd[first_conv].shape[1])
+            self.state_dict()[first_conv][:, :channel] = csd[first_conv][:, :channel]
         if verbose:
             LOGGER.info(f"Transferred {len(csd)}/{len(self.model.state_dict())} items from pretrained weights")