--- a/ultralytics/nn/tasks.py
+++ b/ultralytics/nn/tasks.py
@@ -285,8 +285,11 @@ class BaseModel(torch.nn.Module):
         self.load_state_dict(updated_csd, strict=False)  # load
         first_conv = "model.0.conv.weight"
         if first_conv not in updated_csd:
-            channel = min(self.state_dict()[first_conv].shape[1], csd[first_conv].shape[1])
-            self.state_dict()[first_conv][:, :channel] = csd[first_conv][:, :channel]
+            c1, c2, h, w = self.state_dict()[first_conv].shape
+            cc1, cc2, ch, cw = csd[first_conv].shape
+            if ch == h and cw == w:
+                c1, c2 = min(c1, cc1), min(c2, cc2)
+                self.state_dict()[first_conv][:c1, :c2] = csd[first_conv][:c1, :c2]
         if verbose:
             LOGGER.info(f"Transferred {len(csd)}/{len(self.model.state_dict())} items from pretrained weights")