--- a/ultralytics/nn/tasks.py
+++ b/ultralytics/nn/tasks.py
@@ -285,7 +285,7 @@ class BaseModel(torch.nn.Module):
         self.load_state_dict(updated_csd, strict=False)  # load
         len_updated_csd = len(updated_csd)
         first_conv = "model.0.conv.weight"
-        if first_conv not in updated_csd:
+        if first_conv not in updated_csd:  # mostly used to boost multi-channel training
             c1, c2, h, w = self.state_dict()[first_conv].shape
             cc1, cc2, ch, cw = csd[first_conv].shape
             if ch == h and cw == w: