<!--
Thank you 🙏 for your contribution to [Ultralytics](https://www.ultralytics.com/) 🚀! Your effort in enhancing our repositories is greatly appreciated. To streamline the process and assist us in integrating your Pull Request (PR) effectively, please follow these steps:

1. Check for Existing Contributions: Before submitting, kindly explore existing PRs to ensure your contribution is unique and complementary.
2. Link Related Issues: If your PR addresses an open issue, please link it in your submission. This helps us better understand the context and impact of your contribution.
3. Elaborate Your Changes: Clearly articulate the purpose of your PR. Whether it's a bug fix or a new feature, a detailed description aids in a smoother integration process.
4. Ultralytics Contributor License Agreement (CLA): To uphold the quality and integrity of our project, we require all contributors to sign the CLA. Please confirm your agreement by commenting below:

    I have read the CLA Document and I sign the CLA

For more detailed guidance and best practices on contributing, refer to our ✅ [Contributing Guide](https://docs.ultralytics.com/help/contributing/). Your adherence to these guidelines ensures a faster and more effective review process.
-->


## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Improves model weight loading to better support multi-channel training and provides clearer logging. 🚀

### 📊 Key Changes
- Enhances the weight loading process to handle cases where the first convolutional layer's weights don't match, especially for multi-channel inputs.
- Updates the logging to accurately reflect the number of parameters transferred from pretrained weights.

### 🎯 Purpose & Impact
- Makes it easier to use pretrained models with custom input channels (e.g., more than 3 for RGB), which is useful for advanced use cases like multi-spectral imagery.
- Provides users with clearer feedback on how many weights were successfully loaded, improving transparency and debugging.
- Reduces potential errors or mismatches when adapting models to new data types.