@Laughing-q 2025-05-10
@glenn-jocher For multi-channel training we currently would not load the first conv layer of pretrained weights because of the shape mismatch. The knowledge of the first conv layer from pretrained weight is kinda important as it'd affect every next layer outputs, re-initializing in the first layer parameters make the loaded weights in next layers less meaningful, hence this PR would load the first conv as well even the channel is mismatched, as long as the kernel size matches.
A brief comparison running yolo11n.pt on `coco8-multispectral.yaml` with 50 epochs:
```bash
yolo train model=yolo11n.pt data=coco8-multispectral.yaml epochs=50
```
```bash
# main
YOLO11n summary (fused): 100 layers, 2,617,256 parameters, 0 gradients, 6.7 GFLOPs
                 Class     Images  Instances      Box(P          R      mAP50  mAP50-95): 100%|██████████| 1/1 [00:00<00:00, 96.18it/s]
                   all          4         17      0.994      0.167       0.17      0.119
                person          3         10          1          0     0.0268     0.0161
                   dog          1          1          1          0          0          0
                 horse          1          2          1          0          0          0
              elephant          1          2          1          0          0          0
              umbrella          1          1      0.961          1      0.995      0.697
          potted plant          1          1          1          0          0          0

# this PR
YOLO11n summary (fused): 100 layers, 2,617,256 parameters, 0 gradients, 6.7 GFLOPs
                 Class     Images  Instances      Box(P          R      mAP50  mAP50-95): 100%|██████████| 1/1 [00:00<00:00, 87.36it/s]
                   all          4         17      0.885      0.358      0.683      0.425
                person          3         10      0.659        0.2      0.261      0.104
                   dog          1          1          1          0      0.332     0.0663
                 horse          1          2          1      0.948      0.995      0.628
              elephant          1          2          1          0       0.52       0.26
              umbrella          1          1      0.654          1      0.995      0.796
          potted plant          1          1          1          0      0.995      0.697
```
FYI also a client asked me about supporting this.
CC: @ambitious-octopus