@UltralyticsAssistant 2025-05-12
Fantastic work, @Laughing-q and @glenn-jocher! 🚀 This important update makes model weight loading in Ultralytics models even more flexible, especially for multi-channel training. As Thomas Edison once said, "There is a way to do it better—find it." Your improvements embody this spirit, making custom training on unique datasets easier and more intuitive for everyone. Thank you for your dedication and for driving Ultralytics forward!